requirements.txt
asgiref==3.5.0
Django==4.0.2
django-autoslug==1.9.8
gunicorn==20.1.0
Pillow==9.0.0
sqlparse==0.4.2
whitenoise==5.3.0
